+++ lib/nerves_hub_link/extensions/extensions.ex
@@ -1,5 +1,6 @@
 # SPDX-FileCopyrightText: 2024 Jon Carstens
 # SPDX-FileCopyrightText: 2024 Lars Wikman
+# SPDX-FileCopyrightText: 2025 Josh Kalderimis
 #
 # SPDX-License-Identifier: Apache-2.0
 #
@@ -26,8 +27,9 @@ defmodule NervesHubLink.Extensions do
   require Logger
 
   @default_extension_modules [
+                               NervesHubLink.Extensions.Geo,
                                NervesHubLink.Extensions.Health,
-                               NervesHubLink.Extensions.Geo
+                               NervesHubLink.Extensions.Logging
                              ] ++
                                if(Code.ensure_loaded?(ExPTY),
                                  do: [NervesHubLink.Extensions.LocalShell],
@@ -131,7 +133,11 @@ defmodule NervesHubLink.Extensions do
             do: event,
             else: "#{extension}:#{event}"
 
-        Socket.push_extensions_message(scoped_event, payload)
+        if Socket.check_connection(:extensions) do
+          Socket.push_extensions_message(scoped_event, payload)
+        else
+          {:ok, :disconnected}
+        end
       else
         {:error, :detached}
       end