+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+Added
+  * Add support for a new Logging extension
+
 ## [2.11.1] - 2026-03-23
 
 * Fixed
@@ -71,6 +76,7 @@ This release shouldn't affect your setup unless you are doing fairly unusual con
 
 * Fixed
   * SSL: `cacerts` no longer override `cacertfile` (#321) (Thanks @D4no0)
+
 
 ## [2.7.3] - 2025-04-15